WIP: edited evaluator utils dream
